moved tests to unit tests where applicable
