+++ tests/integration/detection/test_test_case.py
@@ -449,13 +449,3 @@ def test__create__with_locator_collision() -> None:
     test_case = TestCase(test_case_name, images=[TestImage(locator)])
     images = test_case.load_images()
     assert len(images) == 1
-
-
-def test__init__validate_name() -> None:
-    with pytest.raises(ValueError):
-        TestCase("")
-
-
-def test__create__validate_name() -> None:
-    with pytest.raises(ValueError):
-        TestCase.create("")
